Keyboard input and rotation of Blue ricky
気持ち悪い

This is... interesting. This probably has more to do
with the fact that the tutorial I'm following is
in swing and so maybe its not as unusual to have
shared global mutable state but good lord does it
make me shiver. I'm going to refactor this shit SO
hard once I'm done with the tutorial because the
fact that the mino classes are aware of both the
playmanager AND the keyboard inputs coming in
rather than delegation feels really gross.