Mino collision detection
The rotation detection is pretty clever I think.
This code is slightly different from the tutorial in
that the tutorial used 3 for loops while I use one.

I also guarded the active drop counter because we
don't need to add a number up until max integer if
we're not updating something.