Add next Mino behavior
Add a new mino when the current one hits a bottom
collision case, then transfer those blocks to a list
of static blocks that exist.

note that the last collision we did was only in
respect to the walls, not other blocks. So, that's
up next and I fear the tutorial will have me have
the entire array of static blocks become Java
static...