Make it easier to control
Precision movements are easier if you have to
press the button each time you want to move it to
the left or right.

This could be bette implemented with a delay and
accum value like we do the sliding, but for now
this works so that I can play test easier.