Extend InputAdapter for keyboard inputs
This is interesting in that the adapter processes the
events BEFORE we hit the render step, which is a nice
separation fo church and state.
I don't know if I like passing in the keyboard input to
the update method though. It makes some degree of sense,
but I also want to just have the keyboard input be set
for the Mino's as well and have them refer to that state
that has changed.
I thought about maybe setting a keyboard object in the
mino class, that way it would key automatically
updated if it's shared between the two places. But I
don't know. Perhaps static global state for the input
wouldn't be the worst thing in the world since input
IS global to the app in some ways. I'll keep thinking,
but for now I'll commit this since I do like that the
addition of the keyboard configuration class would let
us remap our controls as needed.