Create GameOverScreen
Going to abstract things out into 3. One splash screen/title
screen. Then the playmanager will become the game screen, and
then of course our gameover screen here.

I'm not sure about the input handler. I found a tutorial on
happycoding.io that had some good notes about setting the
input handler to null between screens and such if you didn't
want it to be shared between the screens, but we'll see if I
abstract it out more or not.