Rename PlayManager to GameScreen + implement screen
Since we'll be sharing batch and friends around places we'll
move those up into PictureMino and also have a way to swap the
screens exposed for the screens to use.

A lot of examples have setScreen(new ScreenClass()) type stuff
which is fine in example, but I have to imagine that not being
able to call dispose on the screens when you're done with them
is a great way to leak memory and run into performance problems
in any sort of long-lived application scrub test. So, I figure
I'll do it this way. If we did do the set and forget, we would
probably need to have the screen call .dispose before being
swapped away from.