Add TitleScreen to start
Nothing special, but it feels good to make one
so the game doesn't start the moment the app
boots up.

If I had title specific music I'd play it here.
And if we allowed for settings to be... set we'd
probably link off to it from here with a
selectable button or something similar.