Move settings inputs to input package
Going to rename some I think. Also I'm not
entirely sold on the SettingSetGet being here
since it is more setting screen specific and
not quite an input concern.

Also... If I just made a key class of some kind,
then I might be able to ditch the setter/getter
altogether which would be nice since it's just a
way to avoid lambdas that looked kind of gross.